moved requireauth routes into route dir
