--- src/components/auth/RequireAuth.jsx
@@ -1,17 +1,0 @@
-import { useLocation, Navigate, Outlet } from 'react-router-dom';
-import { useAuth } from '@/hooks/useAuth';
-
-export default function RequireAuth({ requiredRole }) {
-  const { role, user } = useAuth();
-  const location = useLocation();
-
-  if (!user) {
-    return <Navigate to="/login" state={{ from: location }} replace />;
-  }
-
-  if (requiredRole && role !== requiredRole) {
-    return <Navigate to="/unauthorized" replace />;
-  }
-
-  return <Outlet />;
-}
