+++ src/routes/Dashboard/components/StatusColumn/styles.module.css
